+++ adapters/ikea/tradfri_remote_close_open.py
@@ -1,6 +1,8 @@
+import Domoticz
+import json
 from adapters.adapter_with_battery import AdapterWithBattery
 from devices.switch.selector_switch import SelectorSwitch
-import Domoticz
+
 
 class TradfriRemoteCloseOpen(AdapterWithBattery):
     def __init__(self, devices):